requirements.txt
# Flask-SocketIO
flask-socketio==5.3.2  # Ensure you're using the correct version
pandas
openpyxl

# Flask CORS
flask-cors==5.0.1

# Database dependencies
sqlalchemy==2.0.12
psycopg2==2.9.6
python-dotenv==1.0.0
Flask==2.3.2

# Google Cloud SDK and Vertex AI dependencies
google-cloud-aiplatform==1.83.0  # Updated to the latest available version
google-genai==1.5.0

# Additional utilities
requests==2.28.2
beautifulsoup4==4.13.3
lxml==5.3.1
